Use 1 URL function from all Remote ViewControllers Clean out uneccessary duplicated code (URL request functions in all view controllers -> 1 func in twilio.swift)
